+++ frontend/src/app/model/web-socket.model.ts
@@ -1,7 +1,7 @@
 import { createStore, createEvent, createEffect, sample } from "effector";
 
 import { lessonModel } from "@entities/lesson";
-import { LessonStatus } from "@shared/types";
+import type { LessonStatus } from "@shared/types";
 
 // Events
 export const connectWebSocket = createEvent();